Extract empty tests
