+++ sparse_strips/vello_cpu/tests/mod.rs
@@ -8,6 +8,7 @@
 mod basic;
 mod clip;
 mod compose;
+mod empty;
 mod glyph;
 mod issues;
 mod util;
